# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sage Oak App Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user account (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /usage/daily:
    get:
      operationId: getDailyUsage
      tags: [usage]
      summary: Full-history daily student and teacher usage
      parameters:
        - name: startDate
          in: query
          required: false
          schema: { type: string }
        - name: endDate
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Daily rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyUsageRow"
...
  /usage/additional-resources/history:
    get:
      operationId: getAdditionalResourcesHistory
      tags: [usage]
      summary: Per-resource usage across recent snapshot dates
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 60 }
          description: How many recent snapshot dates to include (default 12)
      responses:
        "200":
          description: Usage history per resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceUsageHistory"
  /uploads:
    post:
      operationId: uploadUsageData
      tags: [uploads]
      summary: Upload a batch of Clever export CSVs (admin, multipart form)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadBatch"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /uploads/sftp/sync:
    post:
      operationId: triggerSftpSync
      tags: [uploads]
      summary: Trigger an immediate Clever SFTP report sync (admin)
      responses:
        "200":
          description: Sync summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SftpSyncSummary"
        "400":
          description: SFTP not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "409":
          description: A sync is already running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /uploads/sftp/alerts/{id}/dismiss:
    post:
      operationId: dismissSyncAlert
      tags: [uploads]
      summary: Dismiss a sync failure alert (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "404":
          description: Alert not found or already resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      properties:
        message: { type: string }
      required: [message]
...
    User:
      type: object
      required: [id, email, displayName, role, tags, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        displayName: { type: string }
        role:
          type: string
          enum: [admin, staff]
        tags:
          type: array
          items: { type: string }
        createdAt: { type: string }
...
    UserInput:
      type: object
      required: [email, password, displayName, role]
      properties:
        email: { type: string }
        password: { type: string }
        displayName: { type: string }
        role:
          type: string
          enum: [admin, staff]
        tags:
          type: array
          items: { type: string }
    UserUpdate:
      type: object
      properties:
        displayName: { type: string }
        role:
          type: string
          enum: [admin, staff]
        password: { type: string }
        tags:
          type: array
          items: { type: string }
...
    DailyUsageRow:
      type: object
      required: [date, studentUsers, teacherUsers]
      properties:
        date: { type: string }
        studentUsers:
          type: ["integer", "null"]
        teacherUsers:
          type: ["integer", "null"]
...
    ResourceUsagePoint:
      type: object
      required: [snapshotDate, uniqueUsers, totalAccesses]
      properties:
        snapshotDate: { type: string }
        uniqueUsers: { type: integer }
        totalAccesses: { type: integer }
    ResourceUsageSeries:
      type: object
      required: [link, points]
      properties:
        link: { type: string }
        points:
          type: array
          items:
            $ref: "#/components/schemas/ResourceUsagePoint"
    ResourceUsageHistory:
      type: object
      required: [snapshotDates, resources]
      properties:
        snapshotDates:
          type: array
          items: { type: string }
        resources:
          type: array
          items:
            $ref: "#/components/schemas/ResourceUsageSeries"
    UploadFile:
      type: object
      required: [name, content]
      properties:
        name: { type: string }
        content: { type: string }
    UploadBatch:
      type: object
      required: [files]
      properties:
        files:
          type: array
          items:
            $ref: "#/components/schemas/UploadFile"
    ImportResult:
      type: object
      required: [snapshotDate, filesProcessed, rowsInserted, rowsUpdated, warnings]
      properties:
        snapshotDate: { type: string }
        filesProcessed:
          type: array
          items: { type: string }
        rowsInserted: { type: integer }
        rowsUpdated: { type: integer }
        warnings:
          type: array
          items: { type: string }
...
    SftpSyncSummary:
      type: object
      required: [importedSnapshots, skippedSnapshots, warnings]
      properties:
        importedSnapshots:
          type: array
          items: { type: string }
        skippedSnapshots:
          type: array
          items: { type: string }
        warnings:
          type: array
          items: { type: string }
...