# lib/api-zod/src/generated/types/resourceUsageHistory.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sage Oak App Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceUsageSeries } from './resourceUsageSeries';

export interface ResourceUsageHistory {
  snapshotDates: string[];
  resources: ResourceUsageSeries[];
}