# lib/api-zod/src/generated/types/resourceUsageSeries.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sage Oak App Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceUsagePoint } from './resourceUsagePoint';

export interface ResourceUsageSeries {
  link: string;
  points: ResourceUsagePoint[];
}